Split out Message, add interface
